We now only process messages from the same environment we are running in.

+++ cosmosThrottler/writeToCosmos.js
@@ -3,19 +3,22 @@
 module.exports = async function(_, msg) {
     const parsed = JSON.parse(msg);
     console.log(cosmos);
-    switch(parsed.command)
+    if (parsed.environment == process.env.ENVIRONMENT)
     {
-        case 'add-vertex':
-            await cosmos.createEntryOfKind(parsed.kind, parsed.id, parsed.properties, parsed.edges);
-            break;
-        case 'add-edge':
-            await cosmos.createEdge(parsed.source, parsed.target, parsed.relationship, parsed.properties);
-            break;
-        case 'delete-vertex':
-            await cosmos.deleteEntry(parsed.id, parsed.edgeLabelsToFollow);
-            break;
-        case 'delete-edge':
-            await cosmos.deleteEdge(parsed.id);
-            break;
+        switch(parsed.command)
+        {
+            case 'add-vertex':
+                await cosmos.createEntryOfKind(parsed.kind, parsed.id, parsed.properties, parsed.edges);
+                break;
+            case 'add-edge':
+                await cosmos.createEdge(parsed.source, parsed.target, parsed.relationship, parsed.properties);
+                break;
+            case 'delete-vertex':
+                await cosmos.deleteEntry(parsed.id, parsed.edgeLabelsToFollow);
+                break;
+            case 'delete-edge':
+                await cosmos.deleteEdge(parsed.id);
+                break;
+        }
     }
 };
